- quick hashrate calculations - start updating API info

+++ src/Miningcore/Mining/PoolStats.cs
@@ -26,7 +26,7 @@ namespace Miningcore.Mining
     {
         public DateTime? LastPoolBlockTime { get; set; }
         public int ConnectedMiners { get; set; }
-        public ulong PoolHashrate { get; set; }
-        public int SharesPerSecond { get; set; }
+        public double PoolHashrate { get; set; }
+        public double SharesPerSecond { get; set; }
     }
 }
